proyecto_rag/backend/requirements.txt
pandas
numpy
nltk
spacy
transformers
sentence-transformers
faiss-cpu
streamlit
scikit-learn